+++ apps/posts/models/__init__.py
@@ -1,4 +1,4 @@
-from posts.models.posts_category_code_models import CategoryCode
+from posts.models.posts_category_code_model import CategoryCode
 from posts.models.company_model import Company
 from posts.models.product_model import Product
 from posts.models.posts_product_image_model import ProductImage
